+++ src/zmq.cpp
@@ -859,7 +859,7 @@ int zmq_poll (zmq_pollitem_t *items_, int nitems_, long timeout_)
         ns_.tv_nsec = timeout_ % 1000 * 1000000;
         return nanosleep (&ns_, 0);
 #else
-        return usleep (timeout_ * 1000);
+        return usleep ((unsigned int)timeout_ * 1000);
 #endif
     }
     if (!items_) {
